hanlde errors cart in service-cart

+++ apps/service-cart/src/cart/cart.controller.ts
@@ -20,13 +20,13 @@ export class CartController {
   constructor(private readonly cartService: CartService) {}
 
   @Post()
-  create(@Res() res: Response, @Body() createCartDto: CreateCartDto) {
+  async create(@Res() res: Response, @Body() createCartDto: CreateCartDto) {
     try {
-      this.cartService.create(createCartDto).then((data) => {
-        res.status(HttpStatus.CREATED).json({
-          message: 'Cart created',
-          data: data,
-        });
+      const cart = await this.cartService.create(createCartDto);
+
+      return res.status(HttpStatus.CREATED).json({
+        message: 'Cart created successfully',
+        data: cart,
       });
     } catch (error) {
       res.status(HttpStatus.BAD_REQUEST).json({
@@ -37,13 +37,13 @@ export class CartController {
   }
 
   @Get()
-  findAll(@Res() res: Response) {
+  async findAll(@Res() res: Response) {
     try {
-      this.cartService.findAll().then((data) => {
-        res.status(HttpStatus.OK).json({
-          message: 'Carts found',
-          data: data,
-        });
+      const carts = await this.cartService.findAll();
+
+      return res.status(HttpStatus.OK).json({
+        message: 'Carts found',
+        data: carts,
       });
     } catch (error) {
       res.status(HttpStatus.BAD_REQUEST).json({
@@ -54,7 +54,7 @@ export class CartController {
   }
 
   @Get(':id')
-  findOne(
+  async findOne(
     @Param(
       'id',
       new ParseIntPipe({ errorHttpStatusCode: HttpStatus.NOT_ACCEPTABLE }),
@@ -63,11 +63,10 @@ export class CartController {
     @Res() res: Response,
   ) {
     try {
-      this.cartService.findOne(+id).then((data) => {
-        res.status(HttpStatus.OK).json({
-          message: 'Cart found',
-          data: data,
-        });
+      const cart = await this.cartService.findOne(+id);
+      return res.status(HttpStatus.OK).json({
+        message: 'Cart found',
+        data: cart,
       });
     } catch (error) {
       res.status(HttpStatus.BAD_REQUEST).json({
@@ -78,25 +77,65 @@ export class CartController {
   }
 
   @Patch(':id')
-  update(
+  async update(
     @Param(
       'id',
       new ParseIntPipe({ errorHttpStatusCode: HttpStatus.NOT_ACCEPTABLE }),
     )
     id: number,
     @Body() updateCartDto: UpdateCartDto,
+    @Res() res: Response,
   ) {
-    return this.cartService.update(+id, updateCartDto);
+    try {
+      const check = await this.cartService.cartExists(+id);
+
+      if (!check) {
+        return res.status(HttpStatus.NOT_FOUND).json({
+          message: 'Cart not found',
+          data: null,
+        });
+      }
+      const cart = await this.cartService.update(+id, updateCartDto);
+
+      return res.status(HttpStatus.OK).json({
+        message: 'Cart updated successfully',
+        data: cart,
+      });
+    } catch (error) {
+      res.status(HttpStatus.BAD_REQUEST).json({
+        message: 'Cart not updated',
+        data: error,
+      });
+    }
   }
 
   @Delete(':id')
-  remove(
+  async remove(
     @Param(
       'id',
       new ParseIntPipe({ errorHttpStatusCode: HttpStatus.NOT_ACCEPTABLE }),
     )
     id: number,
+    @Res() res: Response,
   ) {
-    return this.cartService.remove(+id);
+    try {
+      const check = await this.cartService.cartExists(+id);
+      if (!check) {
+        return res.status(HttpStatus.NOT_FOUND).json({
+          message: 'Cart not found',
+          data: null,
+        });
+      }
+      const cart = await this.cartService.remove(+id);
+      return res.status(HttpStatus.OK).json({
+        message: 'Cart deleted successfully',
+        data: cart,
+      });
+    } catch (error) {
+      res.status(HttpStatus.BAD_REQUEST).json({
+        message: 'Cart not deleted',
+        data: error,
+      });
+    }
   }
 }
